Move game client codecs into separate file
